requirements.txt
ag2[openai]==0.10.5
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
azure-core==1.38.0
azure-functions==1.24.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
azure-storage-blob==12.27.1
babel==2.17.0
backrefs==6.1
bcrypt==4.0.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
csscompressor==0.9.5
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
ecdsa==0.19.1
fastapi==0.128.0
feedparser==6.0.12
frozenlist==1.8.0
ghp-import==2.1.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
htmlmin2==0.1.13
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
jsmin==3.0.1
jsonref==1.1.0
Markdown==3.10.1
markdownify==1.2.2
MarkupSafe==3.0.3
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.3
mkdocs-get-deps==0.2.0
mkdocs-git-revision-date-localized-plugin==1.5.0
mkdocs-material==9.7.1
mkdocs-material-extensions==1.3.1
mkdocs-minify-plugin==0.8.0
motor==3.7.1
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
openai==2.14.0
opentelemetry-api==1.39.1
packaging==26.0
paginate==0.5.7
passlib==1.7.4
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==5.29.5
pyasn1==0.6.2
pyautogen==0.10.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.20
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pywin32==311
PyYAML==6.0.3
pyyaml_env_tag==1.1
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
ruff==0.14.14
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.50.0
structlog==25.5.0
tenacity==9.1.2
termcolor==3.3.0
tiktoken==0.12.0
tqdm==4.67.1
tree-sitter==0.25.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
vulture==2.14
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
yarl==1.22.0
zipp==3.23.0
falkordb>=1.0.0